refactor(BA-7353): move the kernel payload into common/interchange and carry only what the manager reads
Review follow-up. `common/data` is where a component keeps its own value objects, so a
payload that crosses a component boundary had no documented home — `dto/` is organized
by target component, and a payload has two sides.

- Add `common/interchange/`, with `AGENTS.md` stating the membership test: one component
  produces it, another consumes it, so it has to survive serialization. Record the split
  between `interchange` / `dto` / `schema` / `data` in `common/KNOWLEDGE.md`, and list the
  package (along with the previously missing `schema/`) in `common/AGENTS.md`.
- Carry only what `update_kernel_status_running()` reads, which is its one consumer:
  `container_id`, `kernel_host`, the two repl ports, `service_ports`, `attached_devices`
  and `allocations`. Dropped `id` (the event already names the kernel), `agent_addr` (the
  event names its source), the legacy stdio ports, and the parts of the resource spec —
  slots, mounts, unified devices, scratch size — that nothing reads.
- Return `list[ResourceSlotEntry]` rather than the legacy `ResourceSlot`; a caller that
  needs the dict form goes through `ResourceSlotEntry.inputs_to_resource_slot()`.
- Name the levels of the allocation mapping (`DeviceAllocation`, `PerDeviceAllocation`)
  instead of nesting three dicts inline, and apply the `DeviceAllocation` alias that
  `agent/resources.py` already defined but did not use.

With the spec no longer travelling whole, `KernelResourceSpecData` / `MountData` /
`to_data()` would only serve the agent's own log, so they are gone and
`to_json_serializable_dict()` is left as it was.
